pyproject.toml

[project]
name = "heli_code_backend"
version = "0.1.0"
description = "A simple FastAPI backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.7",
    "uvicorn[standard]>=0.39.0",
    "numpy>=2.3.5",
    "torch>=2.9.1",
    "torchvision>=0.24.0",
    "opencv-python>=4.8.1.78",
    "onnxruntime>=1.23.0",
    "python-dotenv>=1.1.1",
    "pydantic-settings>=2.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]